requirements.txt
groq
sentence-transformers
scikit-learn
pandas
numpy
python-dotenv
fastapi
joblib
uvicorn